Jack socket breadboard mount.
This is to receive an enclosed chassis 3.5mm jack mount (RS part #106874) whose active pin are to be extended soldering additional wire. The sleeve pin conduit is at an angle so that it exits 1/10 of an inch to the right respective of the tip pin which is at the back.

The other distance between the tip and the sleeve pin should be 5/10 of an inch, making the whole contraption breadboard mountable.